move gpui saving to admin/preferences, remove old gpsessions class from global namespace
